Fix: moved the component files into the components folder for more organization
