+++ components/JourneyCard/JourneyCard.tsx
